refactor: move vector index responses under separate subpackage
