+++ src/momento/requests/__init__.py
@@ -1,7 +1,6 @@
 """Moment client request types."""
 
 from .collection_ttl import CollectionTtl
-from .item import Item
 from .sort_order import SortOrder
 
-__all__ = ["CollectionTtl", "Item", "SortOrder"]
+__all__ = ["CollectionTtl", "SortOrder"]
