+++ src/momento/requests/vector_index/__init__.py
@@ -1,0 +1,3 @@
+from .item import Item
+
+__all__ = ["Item"]
