+++ src/momento/responses/vector_index/__init__.py
@@ -1,0 +1,24 @@
+from .control.create import CreateIndex, CreateIndexResponse
+from .control.delete import DeleteIndex, DeleteIndexResponse
+from .control.list import ListIndexes, ListIndexesResponse
+from .data.search import SearchHit, VectorIndexSearch, VectorIndexSearchResponse
+from .data.upsert_item_batch import (
+    VectorIndexUpsertItemBatch,
+    VectorIndexUpsertItemBatchResponse,
+)
+from .response import VectorIndexResponse
+
+__all__ = [
+    "CreateIndex",
+    "CreateIndexResponse",
+    "DeleteIndex",
+    "DeleteIndexResponse",
+    "ListIndexes",
+    "ListIndexesResponse",
+    "SearchHit",
+    "VectorIndexSearch",
+    "VectorIndexSearchResponse",
+    "VectorIndexUpsertItemBatch",
+    "VectorIndexUpsertItemBatchResponse",
+    "VectorIndexResponse",
+]
