+++ src/momento/responses/vector_index/response.py
@@ -1,0 +1,5 @@
+from ..response import Response
+
+
+class VectorIndexResponse(Response):
+    ...
